requirements.txt
streamlit
google-genai
python-dotenv
chromadb
sentence-transformers
python-dotenv
pydantic